refactor(wa-sqlite): seperate adapter
