+++ packages/treecrdt-wa-sqlite/src/adapter.ts
@@ -1,0 +1,19 @@
+import type { TreecrdtAdapter } from '@treecrdt/interface';
+import { createTreecrdtSqliteAdapter, type SqliteRunner } from '@treecrdt/interface/sqlite';
+import type { MaterializationEvent } from '@treecrdt/interface/engine';
+import { dbGetText } from './sql.js';
+import type { Database } from './types.js';
+
+function createRunner(db: Database): SqliteRunner {
+  return {
+    exec: (sql) => db.exec(sql),
+    getText: (sql, params = []) => dbGetText(db, sql, params ?? []),
+  };
+}
+
+export function createWaSqliteApi(
+  db: Database,
+  opts: { maxBulkOps?: number; onMaterialized?: (event: MaterializationEvent) => void } = {},
+): TreecrdtAdapter {
+  return createTreecrdtSqliteAdapter(createRunner(db), opts);
+}
