+++ packages/treecrdt-wa-sqlite/src/index.ts
@@ -1,9 +1,3 @@
-import type { TreecrdtAdapter } from '@treecrdt/interface';
-import { createTreecrdtSqliteAdapter, type SqliteRunner } from '@treecrdt/interface/sqlite';
-import type { MaterializationEvent } from '@treecrdt/interface/engine';
-import { dbGetText } from './sql.js';
-import type { Database } from './types.js';
-
 export type {
   ClientMode,
   ClientOptions,
@@ -33,16 +27,4 @@ export {
 
 export { CLIENT_CLOSED_ERROR, createTreecrdtClient } from './client.js';
 
-function createRunner(db: Database): SqliteRunner {
-  return {
-    exec: (sql) => db.exec(sql),
-    getText: (sql, params = []) => dbGetText(db, sql, params ?? []),
-  };
-}
-
-export function createWaSqliteApi(
-  db: Database,
-  opts: { maxBulkOps?: number; onMaterialized?: (event: MaterializationEvent) => void } = {},
-): TreecrdtAdapter {
-  return createTreecrdtSqliteAdapter(createRunner(db), opts);
-}
+export { createWaSqliteApi } from './adapter.js';
